+++ src/Mock/MockConnection.ts
@@ -1,35 +1,35 @@
-import { IConnection } from "../IConnection";
-import { DeferredQuery } from "../../Query/DeferredQuery";
-import { IQueryResult } from "../../Query/IQueryResult";
-import { Enumerable } from "../../Enumerable/Enumerable";
-import { SelectExpression } from "../../Queryable/QueryExpression/SelectExpression";
-import { QueryType, IsolationLevel, ColumnGeneration } from "../../Common/Type";
-import { IColumnExpression } from "../../Queryable/QueryExpression/IColumnExpression";
-import { InsertExpression } from "../../Queryable/QueryExpression/InsertExpression";
-import { UpdateExpression } from "../../Queryable/QueryExpression/UpdateExpression";
-import { DeleteExpression } from "../../Queryable/QueryExpression/DeleteExpression";
-import { UpsertExpression } from "../../Queryable/QueryExpression/UpsertExpression";
-import { UUID } from "../../Data/UUID";
-import { TimeSpan } from "../../Data/TimeSpan";
-import { IQuery } from "../../Query/IQuery";
-import { BatchedQuery } from "../../Query/BatchedQuery";
-import { EventHandlerFactory } from "../../Event/EventHandlerFactory";
-import { IEventHandler, IEventDispacher } from "../../Event/IEventHandler";
-import { ValueExpression } from "../../ExpressionBuilder/Expression/ValueExpression";
-import { IExpression } from "../../ExpressionBuilder/Expression/IExpression";
-import { PagingJoinRelation } from "../../Queryable/Interface/PagingJoinRelation";
-import { IncludeRelation } from "../../Queryable/Interface/IncludeRelation";
-import { StringColumnMetaData } from "../../MetaData/StringColumnMetaData";
-import { isNotNull } from "../../Helper/Util";
-import { IntegerColumnMetaData } from "../../MetaData/IntegerColumnMetaData";
-import { InsertIntoExpression } from "../../Queryable/QueryExpression/InsertIntoExpression";
-import { ExpressionExecutor } from "../../ExpressionBuilder/ExpressionExecutor";
-import { SqlTableValueParameterExpression } from "../../ExpressionBuilder/Expression/SqlTableValueParameterExpression";
+import { IConnection } from "../Connection/IConnection";
+import { DeferredQuery } from "../Query/DeferredQuery";
+import { IQueryResult } from "../Query/IQueryResult";
+import { Enumerable } from "../Enumerable/Enumerable";
+import { SelectExpression } from "../Queryable/QueryExpression/SelectExpression";
+import { QueryType, IsolationLevel, ColumnGeneration } from "../Common/Type";
+import { IColumnExpression } from "../Queryable/QueryExpression/IColumnExpression";
+import { InsertExpression } from "../Queryable/QueryExpression/InsertExpression";
+import { UpdateExpression } from "../Queryable/QueryExpression/UpdateExpression";
+import { DeleteExpression } from "../Queryable/QueryExpression/DeleteExpression";
+import { UpsertExpression } from "../Queryable/QueryExpression/UpsertExpression";
+import { UUID } from "../Data/UUID";
+import { TimeSpan } from "../Data/TimeSpan";
+import { IQuery } from "../Query/IQuery";
+import { BatchedQuery } from "../Query/BatchedQuery";
+import { EventHandlerFactory } from "../Event/EventHandlerFactory";
+import { IEventHandler, IEventDispacher } from "../Event/IEventHandler";
+import { ValueExpression } from "../ExpressionBuilder/Expression/ValueExpression";
+import { IExpression } from "../ExpressionBuilder/Expression/IExpression";
+import { PagingJoinRelation } from "../Queryable/Interface/PagingJoinRelation";
+import { IncludeRelation } from "../Queryable/Interface/IncludeRelation";
+import { StringColumnMetaData } from "../MetaData/StringColumnMetaData";
+import { isNotNull } from "../Helper/Util";
+import { IntegerColumnMetaData } from "../MetaData/IntegerColumnMetaData";
+import { InsertIntoExpression } from "../Queryable/QueryExpression/InsertIntoExpression";
+import { ExpressionExecutor } from "../ExpressionBuilder/ExpressionExecutor";
+import { SqlTableValueParameterExpression } from "../ExpressionBuilder/Expression/SqlTableValueParameterExpression";
 
 const charList = ["a", "a", "i", "i", "u", "u", "e", "e", "o", "o", " ", " ", " ", "h", "w", "l", "r", "y"];
 let SelectExpressionType: any;
 (async function () {
-    SelectExpressionType = (await import("../../Queryable/QueryExpression/SelectExpression")).SelectExpression;
+    SelectExpressionType = (await import("../Queryable/QueryExpression/SelectExpression")).SelectExpression;
 })();
 export class MockConnection implements IConnection {
     public deferredQueries: Iterable<DeferredQuery>;
